Move more slow tests to the batch 2
